Refactoring RAM & ROM

+++ RAM/Communication.go
@@ -4,7 +4,7 @@ type AddressBus struct {
 	// The Read/Write Byte indicates if the CPU wants to read from or write to memory
 	Rw byte
 	// actual contents
-	Data uint16
+	Data uint32
 }
 
 type DataBus struct {
